Add annotated text functions and remove NullElement
